LAB6/mnist/requirements.txt
# 1. Install TensorFlow for Apple Silicon (M1/M2)
tensorflow==2.15.0
# 2. Install other packages
flask==3.0.0 
numpy==1.24.3 
pillow==10.1.0 
gunicorn==21.2.0